Refactor: Combine event management modules
Consolidated the functionality of `enhanced-events.js` into `combined-events.js` to create a single source of truth for match event handling.

- Merged the more detailed statistics calculation from `enhanced-events.js` into `combined-events.js`.
- Deleted the now-redundant `js/modules/ui/enhanced-events.js` file.
- Removed the obsolete alias export for `enhancedEventsManager` from `combined-events.js`.
- Cleaned up comments to reflect the changes.
